update the type system to stop implicit narrowing. use as.integer, as.double

+++ data/blocks/tail_expr_basic.R
@@ -1,8 +1,8 @@
 # Test basic tail expression in function
-add_and_triple <- function(x: int, y: int): int {
+add_and_triple <- function(x: double, y: double): double {
     sum <- x + y
     sum * 3
 }
 
-result: int <- add_and_triple(5, 3)
+result: double <- add_and_triple(5, 3)
 print(result)
